stock-trading/requirements.txt
# 量化交易系统依赖
requests>=2.28.0
beautifulsoup4>=4.11.0
pandas>=1.5.0
numpy>=1.23.0
python-dotenv>=1.0.0

